Removed germany_2018 module
